Split server setup code into separate file
